refactorizacion en components y styles, he eliminado campos repetidos dentro de post.
