+++ social-network/src/App.jsx
@@ -1,9 +1,8 @@
 import { useState } from 'react'; 
 import './assets/styles/post.css';  
-import Post from './components/post'; 
+import Post from './components/post/post'; 
 
 function App() {
-  const [count, setCount] = useState(0);
 
   const examplePost = {
     title: 'Título de la Publicación',
