+++ social-network/src/assets/styles/components/post.css
@@ -91,15 +91,6 @@ body {
     color: #bbb;
 }
 
-
-.post__stats {
-    display: flex;
-    justify-content: space-between;
-    font-size: 14px;
-    color: #555;
-    margin: 10px 0; 
-}
-
 .post__actions {
     display: flex;
     justify-content: space-between;
